backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg[binary]==3.2.12
python-dotenv==1.0.1
celery==5.3.6
twilio==9.0.4
Pillow==10.4.0
openai==1.40.2
pypdfium2==4.30.0
